Refactored mail-service. Changed status to message in GenericResponse entity, email.html
