+++ src/main/java/com/ncedu/cheetahtest/mail/service/EmailService.java
@@ -1,7 +1,7 @@
 package com.ncedu.cheetahtest.mail.service;
 
 public interface EmailService {
-    void sendSimpleMessage(String to, String text);
+    void sendSimpleMessage(String to, String text, String subject);
 
-    void sendMessageWithAttachment(String to, String text);
+    void sendMessageWithAttachment(String to, String text, String subject);
 }
